SAS: splitting the sas computation out into it's own package
